+++ bluemira/builders/pf_coil.py
@@ -17,7 +17,8 @@ from bluemira.base.designer import Designer
 from bluemira.base.parameter_frame import Parameter, ParameterFrame
 from bluemira.builders.tools import apply_component_display_options, get_n_sectors
 from bluemira.display.palettes import BLUE_PALETTE
-from bluemira.equilibria.coils import Coil, CoilType
+from bluemira.equilibria.coils import Coil
+from bluemira.equilibria.coils.coil_type import CoilType
 from bluemira.geometry.face import BluemiraFace
 from bluemira.geometry.parameterisations import PictureFrame
 from bluemira.geometry.tools import make_circle, offset_wire, revolve_shape
